+++ src/histo.cpp
@@ -5,14 +5,13 @@
 #include <vector>
 #include <cstdint>
 #include "Lightweight_matrix.h"
-
 #ifdef _OPENMP
 #include <omp.h>
 #endif
 
 using namespace Rcpp;
 
-// calculate squared distance using xy
+// calculate squared distance using xy; keep distance as double
 inline double distance(const double x1, const double y1, const double x2, const double y2)
 {
     const double dx = x2 - x1;
@@ -22,6 +21,7 @@ inline double distance(const double x1, const double y1, const double x2, const 
     return dist;
 }
 
+
 // [[Rcpp::export]]
 Rcpp::IntegerMatrix histo_cpp(
     const Rcpp::NumericMatrix &rs_vals,
@@ -29,23 +29,26 @@ Rcpp::IntegerMatrix histo_cpp(
     const Rcpp::NumericMatrix &samples_xy,
     const double within_km = 1000.0, // radius in kilometers to consider ref points
     const double scale = 100000.0,   // correction scale for CRS; 1 if metric, 100,000 otherwise
-    const double bin_width = 0.05,
+    const float bin_width = 0.05,
     const int bin_num = 650,
     int num_threads = -1)
 {
     // convert all Rcpp matrices to custom C++ matrix [faster computation and avoids OpenMp conflicts]
-    Lightweight_matrix<double> rs(rs_vals);
-    Lightweight_matrix<double> pr(pr_vals);
+    Lightweight_matrix<float> rs(rs_vals);
+    Lightweight_matrix<float> pr(pr_vals);
     Lightweight_matrix<double> samples(samples_xy);
     // define the output matrix;
-    Lightweight_matrix<uint64_t> hist(bin_num, bin_num);
+    Lightweight_matrix<uint64_t> histo_matrix(bin_num, bin_num);
 
     const int nr = rs.nrow();
     const int nc = rs.ncol();
 
+    // the inverse of bin-width
+    const float bwi = 1.0 / bin_width;
+
     // squared distance for faster calculation
     // calculating distance squared is faster; one-time rather than many sqrt
-    const double squared_dist = std::pow(within_km * 1000 / scale, 2);
+    const double squared_dist = std::pow(within_km * 1000.0 / scale, 2.0);
 
     // set the number of threads
     #ifdef _OPENMP
@@ -57,12 +60,12 @@ Rcpp::IntegerMatrix histo_cpp(
     #pragma omp parallel for
     for (int i = 0; i < nr; i++)
     {
-        // second iteration over rows
-        for (int j = 0; j < nr; j++)
+        // get the XY from samples for the first sample
+        const auto xi = samples(i, 0);
+        const auto yi = samples(i, 1);
+        // second iteration over rows skipping double counting, (i,j) is equal to (j,i)
+        for (int j = i + 1; j < nr; j++)
         {
-            // get the XY from samples for the first sample
-            const auto xi = samples(i, 0);
-            const auto yi = samples(i, 1);
             // get the XY from samples for the second sample
             const auto xj = samples(j, 0);
             const auto yj = samples(j, 1);
@@ -73,8 +76,8 @@ Rcpp::IntegerMatrix histo_cpp(
             if (dist < squared_dist)
             {
                 // calculate Manhattan distance
-                double rsdist(0.);
-                double prdist(0.);
+                float rsdist(0.);
+                float prdist(0.);
                 for (int m = 0; m < nc; m++)
                 {
                     rsdist += std::fabs(rs(i, m) - rs(j, m));
@@ -82,27 +85,27 @@ Rcpp::IntegerMatrix histo_cpp(
                 }
 
                 // now update the hist with the distances
-                // calculate the row and column
-                int ii = std::floor(prdist / bin_width); // they are both double
-                int jj = std::floor(rsdist / bin_width);
-                // for now, ignore the values overshooting
+                // calculate the row and column; must be floor to correctly get bin number
+                int jj = std::floor(rsdist * bwi); // they are both float
+                int ii = std::floor(prdist * bwi);
+                // for now, ignore the values overshooting; they'll be mostly noise
                 if (ii < bin_num && jj < bin_num)
                 {
                     #pragma omp atomic
-                    hist(ii, jj)++;
+                    histo_matrix(ii, jj)++;
                 }
             }
         }
     }
 
-    // now transpose the matrix; then revert rows to save it in the correct position
+    // now transpose the matrix; then revert rows to save it in the "correct" position
     Rcpp::IntegerMatrix matrix(bin_num, bin_num);
     for (int i = 0; i < bin_num; ++i)
     {
         int ix = bin_num - 1 - i; // compute the reversed row index directly
         for (int j = 0; j < bin_num; ++j)
         {
-            matrix(i, j) = hist(j, ix); // transpose and reverse in one step
+            matrix(i, j) = histo_matrix(j, ix); // transpose and reverse in one step
         }
     }